fix: migrate to Int primitive type
Notification's progress bar accepts Int progress values only.
Since computation manager sends Float progress values between 0 and 1
to be directly consumed by the UI and increase precision, convert
it to a percentage before passing it to the notification.

+++ core/notifications/src/main/kotlin/com/archsoftware/afoil/core/notifications/Notifier.kt
@@ -3,6 +3,6 @@ package com.archsoftware.afoil.core.notifications
 import android.app.Notification
 
 interface Notifier {
-    fun updateComputationServiceNotification(progress: Float)
+    fun updateComputationServiceNotification(progress: Int)
     fun createComputationServiceNotification(computationName: String?): Notification
 }
